refactored: общий рефакторинг
