+++ src/assets/styles/_vars.scss
@@ -18,3 +18,4 @@ $info-background-color: #c8bff1;
 //  FONTS //
 $default-font: "Amethysta", sans-serif;
 $bold-font: 700;
+$font-size-header: 32px;
